+++ src/Services/NotifierStorageService.php
@@ -25,7 +25,7 @@ final class NotifierStorageService
         $backupDirectory = storage_path('app/private');
         File::ensureDirectoryExists($backupDirectory);
 
-        $filename = 'backup-'.Carbon::now()->format('Y-m-d').'.zip';
+        $filename = 'backup-'.Carbon::now()->format('Y-m-d_H-i-s').'.zip';
         $path = $backupDirectory.'/'.$filename;
 
         NotifierLogger::get()->info('➡️ creating backup file');
@@ -78,7 +78,8 @@ final class NotifierStorageService
                 'php_memory_limit' => ini_get('memory_limit'),
                 'url' => config('notifier.backup_url'),
             ]);
-            NotifierLogger::get()->emergency('❌ END OF SESSION ❌');
+
+            throw $th;
         } finally {
             File::delete($path);
             NotifierLogger::get()->info('➡️ backup file cleaned up');